requirements.txt
paho-mqtt==1.6.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
influxdb-client==1.36.1
flask==2.0.1
requests==2.26.0
plotly==5.3.1 